+++ src/Mobs/Skeleton.h
@@ -18,7 +18,7 @@ public:
 	CLASS_PROTODEF(cSkeleton);
 
 	virtual void GetDrops(cItems & a_Drops, cEntity * a_Killer = NULL) override;
-	virtual void MoveToPosition(const Vector3f & a_Position) override;
+	virtual void MoveToPosition(const Vector3d & a_Position) override;
 	virtual void Attack(float a_Dt) override;
 	virtual void SpawnOn(cClientHandle & a_ClientHandle) override;
